This set of changes deals mostly with getting the nroff and the text versions to have the same page breaking when used.
Issue #172 - This is now mostly fixed.  The only known issues left deal with putting in page break before starting appendixes and getting the PI needLines to be recognized in more locations where it is really needed.  Currently it is respected only during before the first section node in a middle, back and section.  It should also be recognized between section nodes and inside of a list.
Issue #166 - We now do a better (but not perfect) job of mking sure that section headings are not orphaned.  If you have two section headings in a row then the first may still be orphaned.
Issue #114 - If there is not an RFC number then XXXX is used for the RFC number for to internal:/rfc.number - matches v1 behavior
 - Legacy-Id: 1097