Fix tracker issue #296
Replace the space between the series info and the series value with a non-breaking space
Change any slashes in the series value so that there is a non-breaking zero width space following it.

If a URL is placed in the series value - then it is still not going to do correct breaking on this.  However this is not something that people should do.

Legacy-Id: 2007